Create main.js
In this example, we are using the express library to create an API endpoint that exports a user table/collection to a CSV file and downloads it. The csv-string library is used to convert the user table/collection to CSV format, and the fs library is used to create a file and write the CSV data to it. The API sets the content type and attachment header and sends the file data as the response.